+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Instructions: Add a subsection under `[Unreleased]` for additions, fixes, change
 
 ## [Unreleased]
 
+## Added
+
+- You can now build standalone targets as PDF, HTML, or a SCORM archive using `pretext build html -i path/to/ptx/file`.  Output will be placed adjacent to source.
+- You can now build standalone targets from within a project (in case you want to build arbitrary files without creating targets for them).  Just add a target with `standalone="yes"` in the manifest.
+
 ## [2.18.3] - 2025-05-19
 
 Includes updates to core through commit: [638c52a](https://github.com/PreTeXtBook/pretext/commit/638c52a4f65df4371f9b7f907ff709bf4fe8b076)